refactor: ConferenceUiState를 ConferencesUiState로 네이밍 변경
