+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/event/conference/ConferenceFragment.kt
@@ -15,9 +15,9 @@ import com.emmsale.presentation.common.views.filterChipOf
 import com.emmsale.presentation.ui.eventdetail.EventDetailActivity
 import com.emmsale.presentation.ui.main.event.conference.recyclerview.ConferenceRecyclerViewAdapter
 import com.emmsale.presentation.ui.main.event.conference.recyclerview.ConferenceRecyclerViewDivider
-import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceItemUiState
 import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceSelectedFilteringDateOptionUiState
 import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceSelectedFilteringOptionUiState
+import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceUiState
 import com.emmsale.presentation.ui.main.event.conferenceFilter.ConferenceFilterActivity
 import java.time.LocalDate
 
@@ -73,10 +73,10 @@ class ConferenceFragment : BaseFragment<FragmentConferenceBinding>() {
     }
 
     private fun setupEventsObserver() {
-        viewModel.conference.observe(viewLifecycleOwner) { eventsResult ->
+        viewModel.conferences.observe(viewLifecycleOwner) { eventsResult ->
             when {
                 eventsResult.isLoadingConferencesFailed -> requireContext().showToast(getString(R.string.all_data_loading_failed_message))
-                !eventsResult.isLoading -> eventAdapter.submitList(eventsResult.conferenceItems) {
+                !eventsResult.isLoading -> eventAdapter.submitList(eventsResult.conferences) {
                     binding.rvEvents.scrollToPosition(0)
                 }
             }
@@ -120,7 +120,7 @@ class ConferenceFragment : BaseFragment<FragmentConferenceBinding>() {
         binding.btnEventFilter.setOnClickListener { navigateToEventFilter() }
     }
 
-    private fun navigateToEventDetail(event: ConferenceItemUiState) {
+    private fun navigateToEventDetail(event: ConferenceUiState) {
         EventDetailActivity.startActivity(requireContext(), event.id)
     }
