+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/event/conference/recyclerview/ConferenceDiffUtil.kt
@@ -1,18 +1,18 @@
 package com.emmsale.presentation.ui.main.event.conference.recyclerview
 
 import androidx.recyclerview.widget.DiffUtil
-import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceItemUiState
+import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceUiState
 
-object ConferenceDiffUtil : DiffUtil.ItemCallback<ConferenceItemUiState>() {
+object ConferenceDiffUtil : DiffUtil.ItemCallback<ConferenceUiState>() {
     override fun areItemsTheSame(
-        oldItem: ConferenceItemUiState,
-        newItem: ConferenceItemUiState,
+        oldItem: ConferenceUiState,
+        newItem: ConferenceUiState,
     ): Boolean =
         oldItem.id == newItem.id
 
     override fun areContentsTheSame(
-        oldItem: ConferenceItemUiState,
-        newItem: ConferenceItemUiState,
+        oldItem: ConferenceUiState,
+        newItem: ConferenceUiState,
     ): Boolean =
         oldItem == newItem
 }
