+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/event/conference/recyclerview/ConferenceRecyclerViewAdapter.kt
@@ -2,11 +2,11 @@ package com.emmsale.presentation.ui.main.event.conference.recyclerview
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceItemUiState
+import com.emmsale.presentation.ui.main.event.conference.uistate.ConferenceUiState
 
 class ConferenceRecyclerViewAdapter(
-    private val onClickConference: (ConferenceItemUiState) -> Unit,
-) : ListAdapter<ConferenceItemUiState, ConferenceViewHolder>(ConferenceDiffUtil) {
+    private val onClickConference: (ConferenceUiState) -> Unit,
+) : ListAdapter<ConferenceUiState, ConferenceViewHolder>(ConferenceDiffUtil) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ConferenceViewHolder =
         ConferenceViewHolder(parent, onClickConference)
