+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/event/conference/uistate/ConferenceUiState.kt
@@ -1,9 +1,28 @@
 package com.emmsale.presentation.ui.main.event.conference.uistate
 
+import com.emmsale.data.conference.Conference
+import com.emmsale.data.conferenceStatus.ConferenceStatus
+
 data class ConferenceUiState(
-    val conferenceItems: List<ConferenceItemUiState> = emptyList(),
-    val isLoading: Boolean = false,
-    val isLoadingConferencesFailed: Boolean = false,
+    val id: Long,
+    val name: String,
+    val tags: List<String>,
+    val status: String,
+    val posterUrl: String?,
 ) {
-    val conferenceSize: Int = conferenceItems.size
+    companion object {
+        fun from(conference: Conference): ConferenceUiState = ConferenceUiState(
+            id = conference.id,
+            name = conference.name,
+            tags = conference.tags,
+            status = getStatus(conference),
+            posterUrl = conference.posterUrl,
+        )
+
+        private fun getStatus(conference: Conference): String = when (conference.status) {
+            ConferenceStatus.IN_PROGRESS -> "진행중"
+            ConferenceStatus.SCHEDULED -> "D-${conference.dDay}"
+            ConferenceStatus.ENDED -> "마감"
+        }
+    }
 }
