+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/main/event/conference/uistate/ConferencesUiState.kt
@@ -1,0 +1,9 @@
+package com.emmsale.presentation.ui.main.event.conference.uistate
+
+data class ConferencesUiState(
+    val conferences: List<ConferenceUiState> = emptyList(),
+    val isLoading: Boolean = false,
+    val isLoadingConferencesFailed: Boolean = false,
+) {
+    val conferenceSize: Int = conferences.size
+}
