+++ android/2023-emmsale/app/src/main/res/layout/fragment_conference.xml
@@ -37,7 +37,7 @@
             android:layout_height="0dp"
             android:layout_marginStart="6dp"
             android:gravity="center"
-            android:text="@{@string/event_count_format(viewModel.conference.conferenceSize)}"
+            android:text="@{@string/event_count_format(viewModel.conferences.conferenceSize)}"
             android:textColor="@color/primary_color"
             android:textSize="12sp"
             android:textStyle="bold"
@@ -131,7 +131,7 @@
             android:layout_width="0dp"
             android:layout_height="0dp"
             android:indeterminate="true"
-            android:visibility="@{viewModel.conference.loading ? View.VISIBLE : View.GONE}"
+            android:visibility="@{viewModel.conferences.loading ? View.VISIBLE : View.GONE}"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintDimensionRatio="0.15"
             app:layout_constraintEnd_toEndOf="parent"
